Audio generation succsefully implemented

+++ src/audio_core/Note.java
@@ -8,9 +8,9 @@ public class Note {
     public byte velocity; //-128: Velocity 0, 127: Velocity 255
     public byte pan; // -128: Left speaker, 127; Right speaker, 0; Both
     public boolean active; //Define if note is active or not
-    public byte cycle; // TODO: DOCUMENT
+    public int cycle; // TODO: DOCUMENT
 
-    public Note(int note, boolean active, byte velocity,byte cycle){
+    public Note(int note, boolean active, byte velocity,int cycle){
         this.note = note;
         this.active = active;
         this.velocity = velocity;
